Improved time formatting and counting rejected in long

+++ src/cosc561/tsp/strategy/BranchAndBoundPath.java
@@ -13,7 +13,7 @@ public class BranchAndBoundPath extends Strategy {
 	
 	private static final int QUEUE_SIZE = 1000;
 
-	private int rejected;
+	private long rejected;
 
 	private Queue<SparseBranch> pathsInProgress;
